Refactor: Modularize settings and utils classes
